replace a.is_some() && a.unwrap() == b with a == Some(b)
It is good to keep `.unwrap()` calls to a minimum, even if we know that
they will not trigger (because we have checks for that case) because code
may change in a way that breaks the checks or there may be typos that
result in another variable being checked for none-None-ness than the one we
want to read.

Making sure we do not have `.unwrap()`s in the first place makes sure of
that.