cmdline: make the init option a String instead of CString
It is somewhat arbitrary that most members of the `CmdlineOptions` struct
have "normal" types (Mostly `String` and `Option<String>`) but `.init`
does not, simply because its single user expects the value to be of
`CString` type.

This would not be much of a problem if C-String literals (`c"/bin/init"`)
were a stable feature in all rust complier versions we want to use.
But they have in fact only been made stable in Rust 1.77[1] released in
March 2024 and have thus not landed in Rust 1.75 used in Yocto scarthgap.

Make `.init` a normal `String` and do not use C-String literals yet.

[1]: https://releases.rs/docs/1.77.0/

Signed-off-by: Leonard Göhrs <leonard@goehrs.eu>