treewide: refactor mkdir and read_file
mkdir, read_file write_file augment the error case with the path in the
error case. This is useful as we otherwise don't know which location
failed. This commit refactors the functions to take any argument that
coerces into a Path - which is implemented by many types.

mkdir now uses create_dir_all as it is not meant to fail if the
directory already exists.

Signed-off-by: Stefan Kerkmann <s.kerkmann@pengutronix.de>