add REUSE meta-data
For the files mentioned in REUSE.toml (which have been edited by
Pengutronix employees only) the license was changed from GPL-2.0-only to
CC0-1.0. For these files we assume that the threshold of
copyrightability has not been reached. So we do not want to derive GPL
obligations from these snippets.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>