init: refactor to simplify custom init applications
The init code can be separated into 3 stages:

1. The basic setup to make the rest possible. This includes mounting
   special filesystems (/proc, /dev, /sys) and parsing the kernel
   commandline.
2. Mounting the root filesystem including all the prep work, such as
   setting up a dm-verity device etc.
3. Switching to the new rootfs and executing the actuall init program.

In a custom init binary, only the second stage is typically modified, or
extra steps are taken between the stages.

Refactor the whole code into a struct with member functions and organize
the code in a way that makes it easy to add custom code inbetween or
instead of the default stages.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>